+++ Engine/renderer/headers/shape.hpp
@@ -19,8 +19,8 @@ class Shape {
         virtual void setCenter(const Vec2<int>& v) = 0;
         virtual void setXCenter(int x) = 0;
         virtual void setYCenter(int y) = 0;
-        virtual void setXVel(int vx) = 0;
-        virtual void setYVel(int vy) = 0;
+        virtual void setXVel(double vx) = 0;
+        virtual void setYVel(double vy) = 0;
 
         // Getters
         virtual Vec2<int> getCenter() const = 0;
